Change how files are saved
There are a few ways to approach hold:
- Run one item per invocation, skipping the already run items.
- Run all items each invocation, but only expose runnable and unique ones.

The `save!` approach allows multiple ruby versions.
It also opens up the possibility to run multiple different benchmarks
together.

The load and save logic has been consolidated to one place.
Also the serialization has just been tweaked to be a single json object.

For the `save!` option, the file is left around - since it can be run
many times. not totally clean, but I find myself using this a lot.